@@ -9,27 +9,32 @@
     "generate": "nuxt generate",
     "preview": "nuxt preview",
     "lint": "eslint .",
-    "lint:fix": "eslint . --fix"
+    "lint:fix": "eslint . --fix",
+    "typecheck": "nuxt typecheck"
   },
   "devDependencies": {
-    "@nuxt/content": "^2.12.1",
-    "@nuxt/eslint": "^0.3.10",
-    "@nuxt/types": "^2.17.3",
-    "@nuxt/typescript-build": "^3.0.2",
-    "@nuxtjs/tailwindcss": "^6.12.0",
-    "@tailwindcss/typography": "^0.5.13",
-    "eslint": "^9.1.1",
-    "gsap": "^3.12.5",
-    "nuxt": "^3.11.2",
-    "typescript": "^5.4.5"
+    "@nuxt/content": "^3.9.0",
+    "@nuxt/eslint": "^1.11.0",
+    "@nuxtjs/tailwindcss": "^6.14.0",
+    "@tailwindcss/typography": "^0.5.19",
+    "@trivago/prettier-plugin-sort-imports": "^6.0.0",
+    "eslint": "^9.39.1",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.4",
+    "gsap": "^3.13.0",
+    "nuxt": "^4.2.1",
+    "nuxt-svgo": "^4.2.6",
+    "prettier": "^3.7.4",
+    "prettier-plugin-tailwindcss": "^0.7.2",
+    "typescript": "^5.9.3",
+    "vue-tsc": "^3.1.6"
   },
   "dependencies": {
-    "feather-icons": "^4.29.2",
-    "perfect-scrollbar": "^1.5.5",
-    "vue": "^3.4.21",
-    "vue-content-loader": "^2.0.1",
-    "vue-feather": "^2.0.0",
-    "vue-router": "^4.3.0",
-    "vue-toast-notification": "^3.1.2"
+    "better-sqlite3": "^12.5.0",
+    "lucide-vue-next": "^0.556.0",
+    "perfect-scrollbar": "^1.5.6",
+    "vue": "^3.5.25",
+    "vue-router": "^4.6.3",
+    "vue-toast-notification": "^3.1.3"
   }
 }
