@@ -1,7 +1,11 @@
 import withNuxt from './.nuxt/eslint.config.mjs'
+import eslintPluginPrettierRecommended from 'eslint-plugin-prettier/recommended'
 
-export default withNuxt({
-  rules: {
-    'vue/no-v-html': ['off'],
+export default withNuxt(
+  {
+    rules: {
+      'vue/no-v-html': ['off'],
+    },
   },
-})
+  eslintPluginPrettierRecommended,
+)
